Revert "Start moving to strategy objects"
This reverts commit 6a81f31aa95b7914df2bd9df1a5ba7beb9c6c3f8.
